architectures/python/rest_api/hexagonal/flask/requirements.txt
# -r requirements/prod.txt
alembic==0.9.6
bcrypt==3.2.0
click==6.7
coverage==4.4.2
flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.9
Flask-Migrate==2.1.1
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.7.1
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==3.0.3
jsonschema==2.6.0
Mako==1.0.7
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.7.3
pytz==2017.3
selenium==3.8.1
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==0.15.5
pandas==1.2.2
firebase-admin==4.5.2
google-cloud-bigquery==2.9.0
google-cloud-bigquery-storage[pandas,pyarrow]==2.3.0

